refactor: move allocate in slots to hooks
