+++ packages/@lwc/engine-core/src/framework/hooks.ts
@@ -4,17 +4,21 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { ArrayFilter, ArrayJoin, assert, isArray, isNull, isUndefined, keys } from '@lwc/shared';
+import {
+    ArrayFilter,
+    ArrayJoin,
+    ArrayPush,
+    assert,
+    create,
+    isArray,
+    isFalse,
+    isNull,
+    isUndefined,
+    keys,
+} from '@lwc/shared';
 import { getClassList, setText, getAttribute, remove, insert } from '../renderer';
 import { EmptyArray, parseStyleText } from './utils';
-import {
-    createVM,
-    allocateInSlot,
-    getAssociatedVMIfPresent,
-    VM,
-    ShadowMode,
-    RenderMode,
-} from './vm';
+import { createVM, getAssociatedVMIfPresent, VM, ShadowMode, RenderMode } from './vm';
 import { VNode, VCustomElement, VElement, VNodes } from '../3rdparty/snabbdom/types';
 import modEvents from './modules/events';
 import modAttrs from './modules/attrs';
@@ -26,6 +30,7 @@ import modStaticStyle from './modules/static-style-attr';
 import { updateDynamicChildren, updateStaticChildren } from '../3rdparty/snabbdom/snabbdom';
 import { patchElementWithRestrictions, unlockDomMutation, lockDomMutation } from './restrictions';
 import { getComponentInternalDef } from './def';
+import { markComponentAsDirty } from './component';
 import { logError } from '../shared/logger';
 
 function observeElementChildNodes(elm: Element) {
@@ -431,6 +436,55 @@ export function removeElmHook(vnode: VElement) {
         const ch = children[j];
         if (!isNull(ch)) {
             ch.hook.remove(ch, elm!);
+        }
+    }
+}
+
+// slow path routine
+// NOTE: we should probably more this routine to the synthetic shadow folder
+// and get the allocation to be cached by in the elm instead of in the VM
+function allocateInSlot(vm: VM, children: VNodes) {
+    const { cmpSlots: oldSlots } = vm;
+    const cmpSlots = (vm.cmpSlots = create(null));
+    for (let i = 0, len = children.length; i < len; i += 1) {
+        const vnode = children[i];
+        if (isNull(vnode)) {
+            continue;
+        }
+        const { data } = vnode;
+        const slotName = ((data.attrs && data.attrs.slot) || '') as string;
+        const vnodes: VNodes = (cmpSlots[slotName] = cmpSlots[slotName] || []);
+        // re-keying the vnodes is necessary to avoid conflicts with default content for the slot
+        // which might have similar keys. Each vnode will always have a key that
+        // starts with a numeric character from compiler. In this case, we add a unique
+        // notation for slotted vnodes keys, e.g.: `@foo:1:1`
+        if (!isUndefined(vnode.key)) {
+            vnode.key = `@${slotName}:${vnode.key}`;
+        }
+        ArrayPush.call(vnodes, vnode);
+    }
+    if (isFalse(vm.isDirty)) {
+        // We need to determine if the old allocation is really different from the new one
+        // and mark the vm as dirty
+        const oldKeys = keys(oldSlots);
+        if (oldKeys.length !== keys(cmpSlots).length) {
+            markComponentAsDirty(vm);
+            return;
+        }
+        for (let i = 0, len = oldKeys.length; i < len; i += 1) {
+            const key = oldKeys[i];
+            if (isUndefined(cmpSlots[key]) || oldSlots[key].length !== cmpSlots[key].length) {
+                markComponentAsDirty(vm);
+                return;
+            }
+            const oldVNodes = oldSlots[key];
+            const vnodes = cmpSlots[key];
+            for (let j = 0, a = cmpSlots[key].length; j < a; j += 1) {
+                if (oldVNodes[j] !== vnodes[j]) {
+                    markComponentAsDirty(vm);
+                    return;
+                }
+            }
         }
     }
 }
